refactor: merge orchestrator files and rename phase reporters
- Merged orchestrator.py and orchestrator_refactored.py into single file
- Added --use-refactored flag to switch between original and JSON-based systems
- Renamed phase_x_refactored.py files to phase_x_reporter.py for clarity
- Fixed function signatures to return proper tuples
- Fixed missing Path import in phase_b_reporter.py
- Updated meson.build to reflect file changes
- Removed redundant orchestrator_refactored.py file
- Both original (text-based) and refactored (JSON-based) systems work correctly
- All validation tests passing

The merged orchestrator maintains full backward compatibility while providing
structured JSON reporting capability when --use-refactored flag is used.
