+++ sample_json_reports/renamed_and_extra.json
@@ -1,6 +1,6 @@
 {
   "schema_version": "1.0.0",
-  "timestamp": "2025-08-26T12:54:30.099485",
+  "timestamp": "2025-08-26T13:38:25.507221",
   "phases_completed": [
     "A"
   ],
@@ -13,7 +13,7 @@
   "phases": {
     "A": {
       "completed": true,
-      "timestamp": "2025-08-26T12:54:30.099517",
+      "timestamp": "2025-08-26T13:38:25.507257",
       "errors": [
         {
           "phase": "A",
@@ -31,7 +31,7 @@
             "Remove this parameter from your configuration",
             "Switch to dev mode if this parameter is needed"
           ],
-          "timestamp": "2025-08-26T12:54:30.099508"
+          "timestamp": "2025-08-26T13:38:25.507247"
         }
       ],
       "warnings": [],
@@ -48,7 +48,7 @@
             "auto_corrected": true
           },
           "suggestions": [],
-          "timestamp": "2025-08-26T12:54:30.099492"
+          "timestamp": "2025-08-26T13:38:25.507230"
         }
       ],
       "results": {
@@ -77,7 +77,7 @@
         "Remove this parameter from your configuration",
         "Switch to dev mode if this parameter is needed"
       ],
-      "timestamp": "2025-08-26T12:54:30.099508"
+      "timestamp": "2025-08-26T13:38:25.507247"
     }
   ],
   "warnings": [],
@@ -94,13 +94,13 @@
         "auto_corrected": true
       },
       "suggestions": [],
-      "timestamp": "2025-08-26T12:54:30.099492"
+      "timestamp": "2025-08-26T13:38:25.507230"
     }
   ],
   "suggestions": [],
   "metadata": {
     "phase": "A",
     "mode": "public",
-    "report_file": "/var/folders/6r/lhkvk3m14ls587p81nfcy0k40000gp/T/tmp_2cs8_io/report.txt"
+    "report_file": "/var/folders/6r/lhkvk3m14ls587p81nfcy0k40000gp/T/tmpcsgk394q/report.txt"
   }
 }
