+++ test/validation_json_reporting/test_phase_a_integration.py
@@ -9,7 +9,7 @@ from pathlib import Path
 import pytest
 import yaml
 
-from supy.data_model.validation.pipeline.phase_a_refactored import (
+from supy.data_model.validation.pipeline.phase_a_reporter import (
     annotate_missing_parameters_refactored,
     build_phase_a_reporter,
 )
