+++ test_json_validation.py
@@ -9,7 +9,7 @@ import tempfile
 from pathlib import Path
 import yaml
 
-from supy.data_model.validation.pipeline.phase_a_refactored import (
+from supy.data_model.validation.pipeline.phase_a_reporter import (
     annotate_missing_parameters_refactored,
 )
